PLAT-6803 - Initial work - remove hc http calls from hc classes
