+++ integration-healthcheck/src/main/java/org/symphonyoss/integration/healthcheck/services/ServiceHealthIndicator.java
@@ -77,21 +77,9 @@ public abstract class ServiceHealthIndicator implements HealthIndicator {
    */
   private static final String SNAPSHOT_VERSION = "-SNAPSHOT";
 
-  /**
-   * Cache period (in seconds) to retrive the service information.
-   */
-  private static final int SERVICE_CACHE_PERIOD_SECS = 20;
-
-  /**
-   * Lock waiting period (in seconds)
-   */
-  private static final int LOCK_PERIOD_SECS = 5;
 
   @Autowired
   protected IntegrationProperties properties;
-
-  @Autowired
-  private AuthenticationProxy authenticationProxy;
 
   @Autowired
   protected ApplicationEventPublisher publisher;
@@ -103,76 +91,28 @@ public abstract class ServiceHealthIndicator implements HealthIndicator {
    * Cache for the service information.
    */
   private IntegrationBridgeService serviceInfoCache;
+  private String serviceName;
+
+  public ServiceHealthIndicator(String serviceName) {
+    super();
+    this.serviceName = serviceName;
+  }
 
   /**
    * Current version
    */
   protected String currentVersion;
 
-  /**
-   * Lock to avoid concurrent execution
-   */
-  private Lock lock = new ReentrantLock();
-
-  /**
-   * Timestamp for the latest execution
-   */
-  private long lastExecution = 0;
-
-  /**
-   * Trigger the health check process on the required service defined in the event.
-   * @param event Health check event
-   */
-  @EventListener
-  public void handleHealthCheckEvent(HealthCheckEventData event) {
-    String serviceName = mountUserFriendlyServiceName();
-
-    LOG.info("Handle health-check event. Service name: {}", serviceName);
-
-    if (serviceName.equals(event.getServiceName())) {
-      retrieveServiceInfo();
-    }
-  }
-
   @Override
   public Health health() {
-    String serviceName = mountUserFriendlyServiceName();
+    String serviceName = this.serviceName;
 
     try {
-      boolean locked = lock.tryLock(LOCK_PERIOD_SECS, TimeUnit.SECONDS);
-
-      if (!locked) {
-        IntegrationBridgeService service = retrieveServiceInfo();
-        return reportServiceHealth(serviceName, service);
-      }
-
-      long nextExecution = lastExecution + TimeUnit.SECONDS.toMillis(SERVICE_CACHE_PERIOD_SECS);
-
-      if (System.currentTimeMillis() > nextExecution) {
-        serviceInfoCache = retrieveServiceInfo();
-        lastExecution = System.currentTimeMillis();
-      }
-
       return reportServiceHealth(serviceName, serviceInfoCache);
     } catch (Exception e) {
       LOG.error(logMessageSource.getMessage(CACHE_IS_NOT_LOADED, serviceName), e);
       return Health.unknown().build();
-    } finally {
-      lock.unlock();
     }
-  }
-
-  /**
-   * Returns the friendly name when it exists, or else, return the toString method for the
-   * Service Name
-   * @return
-   */
-  protected String mountUserFriendlyServiceName() {
-    String friendlyServiceName = getFriendlyServiceName();
-    if (StringUtils.isEmpty(friendlyServiceName)) {
-      friendlyServiceName = getServiceName().toString();
-    }
-    return friendlyServiceName;
   }
 
   /**
@@ -189,37 +129,6 @@ public abstract class ServiceHealthIndicator implements HealthIndicator {
     return Health.status(service.getConnectivity())
         .withDetail(serviceName, service)
         .build();
-  }
-
-  /**
-   * Retrieves the service information like connectivity, current version, and compatibility.
-   * @return Service information
-   */
-  private IntegrationBridgeService retrieveServiceInfo() {
-    String serviceName = mountUserFriendlyServiceName();
-    LOG.info("Retrieve service info: {}", serviceName);
-
-    IntegrationBridgeService service =
-        new IntegrationBridgeService(getMinVersion(), getServiceBaseUrl());
-
-    String healthCheckUrl = getHealthCheckUrl();
-    String healthResponse = getHealthResponse(healthCheckUrl);
-
-    if (healthResponse == null) {
-      service.setConnectivity(Status.DOWN);
-    } else {
-      handleHealthResponse(service, healthResponse);
-
-      String version = retrieveCurrentVersion(healthResponse);
-
-      if (StringUtils.isNotEmpty(version)) {
-        fireUpdatedServiceVersionEvent(version);
-      }
-
-      service.setCurrentVersion(version);
-    }
-
-    return service;
   }
 
   /**
@@ -242,87 +151,6 @@ public abstract class ServiceHealthIndicator implements HealthIndicator {
     }
 
     return version.replace(SNAPSHOT_VERSION, StringUtils.EMPTY);
-  }
-
-  /**
-   * Raise an updated service version event.
-   * @param version Service version
-   */
-  protected void fireUpdatedServiceVersionEvent(String version) {
-    String oldSemanticVersion = getSemanticVersion(currentVersion);
-    String newSemanticVersion = getSemanticVersion(version);
-
-    ServiceVersionUpdatedEventData event =
-        new ServiceVersionUpdatedEventData(mountUserFriendlyServiceName(), oldSemanticVersion,
-            newSemanticVersion);
-    this.currentVersion = version;
-
-    publisher.publishEvent(event);
-  }
-
-  /**
-   * Hits the built URL to the corresponding service.
-   * @param healthCheckUrl Health-check URL
-   * @return Service health check response.
-   */
-  protected String getHealthResponse(String healthCheckUrl) {
-    Client client = getHttpClient();
-
-    if (client == null) {
-      LOG.warn("There is no HTTP client available to perform the health check");
-      return null;
-    }
-
-    Response response = null;
-    try {
-      HttpClientConfig timeouts = properties.getHttpClientConfig();
-      Invocation.Builder invocationBuilder = client.target(healthCheckUrl)
-          .property(ClientProperties.CONNECT_TIMEOUT, timeouts.getConnectTimeout())
-          .property(ClientProperties.READ_TIMEOUT, timeouts.getReadTimeout())
-          .request()
-          .accept(MediaType.APPLICATION_JSON_TYPE);
-
-      LOG.info("Health Check URL: " + healthCheckUrl);
-      response = invocationBuilder.get();
-
-      return retrieveHealthResponse(response);
-    } catch (ProcessingException e) {
-      LOG.error(
-          logMessageSource.getMessage(PROCESSING_EXCEPTION, getHealthCheckUrl(), e.getMessage()),
-          e);
-      return null;
-    } finally {
-      if (response != null) {
-        response.close();
-      }
-    }
-  }
-
-  protected String retrieveHealthResponse(Response response) {
-    Response.Status status = Response.Status.fromStatusCode(response.getStatus());
-    return OK.equals(status) ? response.readEntity(String.class) : null;
-  }
-
-  /**
-   * Gets the HTTP client to be used on health checks, looking on the YAML file for a
-   * integration already been provisioned.
-   * @return the user name.
-   */
-  private Client getHttpClient() {
-
-    for (Application app : this.properties.getApplications().values()) {
-      if (StringUtils.isEmpty(app.getComponent())) {
-        continue;
-      }
-
-      try {
-        return authenticationProxy.httpClientForUser(app.getComponent(), getServiceName());
-      } catch (UnregisteredUserAuthException e) {
-        LOG.warn(logMessageSource.getMessage(UNREGISTERED_USER, app.getComponent()));
-      }
-    }
-
-    return null;
   }
 
   /**
